alert_status/requirements.txt
python-dotenv>=0.21.0
Telethon>=1.26.0
requests>=2.28.1
pytz>=2022.7
psycopg2>=2.9.5